requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
bcrypt>=3.2
pyjwt>=1.7.1
django-cors-headers>=3.5.0
python-dotenv>=0.14.0
django-mailgun>=0.9.1
pyrebase>=3.0.27